Fixed a serious memory leak in the entity radar

+++ src/me/woder/bot/SlotHandler.java
@@ -24,7 +24,7 @@ public class SlotHandler {
                   in.readFully(var2, 0, next);
              }*/
          }
-         return new Slot(slotnum, bid, count, damage, (short)0);
+         return new Slot(slotnum, bid, count, damage, (byte)0);
      }
 
 }
